+++ zenodot-samples/src/dd/kms/zenodot/samples/DynamicTypingSample.java
@@ -4,7 +4,7 @@ import dd.kms.zenodot.api.ExpressionParser;
 import dd.kms.zenodot.api.ParseException;
 import dd.kms.zenodot.api.Parsers;
 import dd.kms.zenodot.api.settings.ParserSettings;
-import dd.kms.zenodot.api.settings.ParserSettingsUtils;
+import dd.kms.zenodot.api.settings.ParserSettingsBuilder;
 import dd.kms.zenodot.api.wrappers.InfoProvider;
 import dd.kms.zenodot.api.wrappers.ObjectInfo;
 
@@ -18,7 +18,7 @@ public class DynamicTypingSample
 	public static void main(String[] args) throws ParseException {
 		TestClass testInstance = new TestClass();
 
-		ParserSettings settings = ParserSettingsUtils.createBuilder()
+		ParserSettings settings = ParserSettingsBuilder.create()
 			.enableDynamicTyping(true)
 			.build();
 		String expression = "getObject().length()";