+++ zenodot-samples/src/dd/kms/zenodot/samples/SimpleExpressionSample.java
@@ -4,7 +4,7 @@ import dd.kms.zenodot.api.ExpressionParser;
 import dd.kms.zenodot.api.ParseException;
 import dd.kms.zenodot.api.Parsers;
 import dd.kms.zenodot.api.settings.ParserSettings;
-import dd.kms.zenodot.api.settings.ParserSettingsUtils;
+import dd.kms.zenodot.api.settings.ParserSettingsBuilder;
 import dd.kms.zenodot.api.wrappers.InfoProvider;
 
 /**
@@ -14,7 +14,7 @@ import dd.kms.zenodot.api.wrappers.InfoProvider;
 public class SimpleExpressionSample
 {
 	public static void main(String[] args) throws ParseException {
-		ParserSettings settings = ParserSettingsUtils.createBuilder().build();
+		ParserSettings settings = ParserSettingsBuilder.create().build();
 		String expression = "Math.max(new int[]{ 2+3, 123/3, 1 << 4 }[1], (int) Math.round(2.718E2))";
 		ExpressionParser parser = Parsers.createExpressionParser(settings);
 		System.out.println("Result: " + parser.evaluate(expression, InfoProvider.NULL_LITERAL).getObject());