+++ zenodot-tests/test/dd/kms/zenodot/tests/common/AbstractTestExecutor.java
@@ -6,7 +6,6 @@ import dd.kms.zenodot.api.debug.ParserConsoleLogger;
 import dd.kms.zenodot.api.debug.ParserLogger;
 import dd.kms.zenodot.api.settings.ObjectTreeNode;
 import dd.kms.zenodot.api.settings.ParserSettingsBuilder;
-import dd.kms.zenodot.api.settings.ParserSettingsUtils;
 import dd.kms.zenodot.api.settings.Variable;
 import dd.kms.zenodot.impl.debug.ParserLoggers;
 import org.junit.Assert;
@@ -18,7 +17,7 @@ public class AbstractTestExecutor<T extends AbstractTestExecutor>
 	protected static final boolean	SKIP_UNSTABLE_TESTS	= "true".equalsIgnoreCase(System.getProperty("skipUnstableTests"));
 
 	protected final Object					testInstance;
-	protected final ParserSettingsBuilder	settingsBuilder			= ParserSettingsUtils.createBuilder().minimumAccessModifier(AccessModifier.PRIVATE);
+	protected final ParserSettingsBuilder	settingsBuilder			= ParserSettingsBuilder.create().minimumAccessModifier(AccessModifier.PRIVATE);
 
 	private boolean							stopAtError				= false;
 	private boolean							printLogEntriesAtError	= false;