+++ zenodot-tests/test/dd/kms/zenodot/tests/evaluationTests/MethodTestSideEffect.java
@@ -3,7 +3,7 @@ package dd.kms.zenodot.tests.evaluationTests;
 import dd.kms.zenodot.api.ParseException;
 import dd.kms.zenodot.api.Parsers;
 import dd.kms.zenodot.api.settings.ParserSettings;
-import dd.kms.zenodot.api.settings.ParserSettingsUtils;
+import dd.kms.zenodot.api.settings.ParserSettingsBuilder;
 import dd.kms.zenodot.api.wrappers.InfoProvider;
 import dd.kms.zenodot.api.wrappers.ObjectInfo;
 import org.junit.Assert;
@@ -15,7 +15,7 @@ public class MethodTestSideEffect
 	public void testSideEffect() {
 		TestClass testInstance = new TestClass();
 		String expression = "f(g(), s)";
-		ParserSettings parserSettings = ParserSettingsUtils.createBuilder().enableDynamicTyping(false).build();
+		ParserSettings parserSettings = ParserSettingsBuilder.create().enableDynamicTyping(false).build();
 		try {
 			ObjectInfo thisValue = InfoProvider.createObjectInfo(testInstance);
 			Parsers.createExpressionParser(parserSettings).evaluate(expression, thisValue);